+++ packages/theme/CHANGELOG.md
@@ -4,8 +4,11 @@
 
 ### Patch Changes
 
+- chore: RSS 生成支持配置多个
+- chore: 分离 shared 方法库
 - Updated dependencies
   - vitepress-plugin-pagefind@0.4.2
+  - vitepress-plugin-rss@0.2.7
 
 ## 0.4.3